requirements.txt
tkinterx
pillow
PyMuPDF
reportlab
tiktoken
verovio
accelerate
ebooklib
pytesseract
paddlepaddle
paddleocr
einops
transformers
requests
supervision
ultralytics
opencv-python-headless
numpy
